回退编译宏隔离

+++ mc2/allto_allv_grouped_mat_mul/op_api/aclnn_allto_allv_quant_grouped_mat_mul.cpp
@@ -114,11 +114,11 @@ static bool CheckNotNull(const aclTensor *gmmX, const aclTensor *gmmWeight, cons
         OP_LOGE(ACLNN_ERR_PARAM_INVALID, "gmmWeightScale should not be null.");
         return false;
     }
-    if (gmmXQuantMode != QuantModeType::PERTENSOR_QUANT) {
+    if (gmmXQuantMode != static_cast<int64_t>(QuantModeType::PERTENSOR_QUANT)) {
         OP_LOGE(ACLNN_ERR_PARAM_INVALID, "gmmXQuantMode should be 1.");
         return false;
     }
-    if (gmmWeightQuantMode != QuantModeType::PERTENSOR_QUANT) {
+    if (gmmWeightQuantMode != static_cast<int64_t>(QuantModeType::PERTENSOR_QUANT)) {
         OP_LOGE(ACLNN_ERR_PARAM_INVALID, "gmmWeightQuantMode should be 1.");
         return false;
     }
